+++ src/tests/store.test.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux';
-import rootReducer from '../../reducers';
+import rootReducer from '../reducers';
 
 const store = createStore(rootReducer);
